[TWINFACES-324] refactor: rename files and folders for factory-pipeline entity
