+++ src/app/workspace/pipelines/page.tsx
@@ -1,4 +1,4 @@
-import { FactoryPipelines } from "@/screens/factoryPipelines";
+import { FactoryPipelines } from "@/screens/factory-pipelines";
 
 export default function PipelinesPage() {
   return <FactoryPipelines />;
