+++ src/entities/factory-pipeline/api/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-search-v1";
