+++ src/entities/factory-pipeline/api/types.ts
