+++ src/entities/factory-pipeline/index.ts
