+++ src/entities/factory-pipeline/libs/hooks/index.ts
