--- src/screens/factoryPipelines/index.ts
@@ -1,1 +1,0 @@
-export * from "./factoryPipelines";
